[#429] Add sample allocation, borrowing and retrieval to OutgoingConnections
